move _MultivarMixin to pre._base.py
