+++ src/opinf/pre/_shiftscale.py
@@ -12,8 +12,7 @@ import numpy as np
 
 from ..errors import LoadfileFormatError
 from ..utils import hdf5_savehandle, hdf5_loadhandle
-from .._multivar import _MultivarMixin
-from ._base import _BaseTransformer
+from ._base import _BaseTransformer, _MultivarMixin
 
 
 # Functional paradigm =========================================================
